Implement public function to be able to consume all TargetTypes
